+++ src/main/kotlin/dev/kingtux/tms/keybinding/SkinLayerKeyBinding.kt
@@ -18,8 +18,7 @@ class SkinLayerKeyBinding(
     override fun onPressed() {
         val client = MinecraftClient.getInstance()
         if (client.player == null) return
-        // TODO: Figure out how to do this in 1.21.3
-        //client.options.togglePlayerModelPart(playerModelPart, !client.options.isPlayerModelPartEnabled(playerModelPart))
+        client.options.setPlayerModelPart(playerModelPart, !client.options.isPlayerModelPartEnabled(playerModelPart))
         sendToggleMessage(
             client.player!!,
             client.options.isPlayerModelPartEnabled(playerModelPart),