+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 ## [0.0.6] (unreleased)
 - Updated to 1.21.3
+- `togglePlayerModelPart` renamed to `setPlayerModelPart`
 
 ## [0.0.5] (2024-08-21)
 - Removed ` dev.kingtux.tms.api.PriorityKeyBinding` to use Amecs API PriorityKeyBinding