+++ CHANGELOG.md
@@ -4,12 +4,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
-## [0.0.6] (unreleased)
+## [0.0.6] (2024-10-27)
 - Updated to 1.21.3
 - `togglePlayerModelPart` renamed to `setPlayerModelPart`
 - Dropped Mouse Keybinding Support
 - Fixed Crash due to Mixin Issue
-- 
+
 
 ## [0.0.5] (2024-08-21)
 - Removed ` dev.kingtux.tms.api.PriorityKeyBinding` to use Amecs API PriorityKeyBinding